refactor: add variables.nix modules
this refactor abstracts away the `hosts/${hostname}/variables.nix` file
so that the variables there are now stored in `config.variables`. this
is done with the intention of enabling future refactors in which that
file is split up into smaller files, should there be need for a larger
amount of modularity.