major refactor: add more theming options
this refactor is done primarily for the purpose of making it easier to
swap themes using stylix. it adds a theme option to the host variables,
and extracts the base16 palette definitions to
modules/core/palettes.nix. in this file, more palettes can be easily
added, and the accent colors can be easily changed.

this also comes with a flake update.